% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_expressions}
\alias{combine_expressions}
\title{Combine multiple expressions into a single expression}
\usage{
combine_expressions(..., sep=' ')
}
\arguments{
\item{...}{multiple text, expressions or calls created with bquote()}
\item{sep}{item separator. character. space(default), blank or comma}
}
\value{
the given expressions combined into a single expression
}
\description{
Combine multiple expressions into a single expression
}
