% Edited by hand by mike
\name{funcs}
\alias{funcs}
\alias{plus}
\alias{\%+\%}
\alias{minus}
\alias{\%-\%}
\alias{juxtapose}
\alias{\%*\%}
\alias{forwardslash}
\alias{\%/\%}
\alias{plusminus}
\alias{\%+-\%}
\alias{divide}
\alias{\%//\%}
\alias{times}
\alias{\%x\%}
\alias{cdot}
\alias{\%.\%}
\alias{neg}
\alias{pos}
\alias{subscript}
\alias{\%subscript\%}
\alias{\%_\%}
\alias{superscript}
\alias{\%superscript\%}
\alias{\%^\%}
\alias{sqrt_}
\alias{eq}
\alias{\%eq\%}
\alias{neq}
\alias{\%neq\%}
\alias{lt}
\alias{\%<<\%}
\alias{le}
\alias{\%<=\%}
\alias{gt}
\alias{\%>>\%}
\alias{ge}
\alias{\%>=\%}
\alias{approximate}
\alias{\%~~\%}
\alias{congruent}
\alias{\%=~\%}
\alias{defined}
\alias{\%==\%}
\alias{proportional}
\alias{\%prop\%}
\alias{distributed}
\alias{\%~\%}
\alias{not}
\alias{plain}
\alias{bold}
\alias{italic}
\alias{bolditalic}
\alias{underline}
\alias{symbol}
\alias{subset}
\alias{\%subset\%}
\alias{subseteq}
\alias{\%subseteq\%}
\alias{notsubset}
\alias{\%notsubset\%}
\alias{supset}
\alias{\%supset\%}
\alias{supseteq}
\alias{\%supseteq\%}
\alias{in_}
\alias{\%in_\%}
\alias{notin}
\alias{\%notin\%}
\alias{hat}
\alias{tilde}
\alias{ring}
\alias{bar}
\alias{widehat}
\alias{widetilde}
\alias{double_arrow}
\alias{\%<->\%}
\alias{right_arrow}
\alias{\%->\%}
\alias{left_arrow}
\alias{\%<-\%}
\alias{up_arrow}
\alias{\%up\%}
\alias{down_arrow}
\alias{\%down\%}
\alias{double_double_arrow}
\alias{\%<=>\%}
\alias{double_right_arrow}
\alias{\%=>>\%}
\alias{double_left_arrow}
\alias{\%<<=\%}
\alias{double_up_arrow}
\alias{\%dblup\%}
\alias{double_down_arrow}
\alias{\%dbldown\%}
\alias{displaystyle}
\alias{textstyle}
\alias{scriptstyle}
\alias{scriptscriptstyle}
\alias{phantom}
\alias{frac}
\alias{\%frac\%}
\alias{over}
\alias{\%over\%}
\alias{atop}
\alias{\%atop\%}
\alias{space}
\alias{\%space\%}
\alias{space2}
\alias{\%space2\%}
\alias{space3}
\alias{\%space3\%}
\alias{space4}
\alias{\%space4\%}
\alias{sum_}
\alias{prod_}
\alias{integral}
\alias{union_}
\alias{intersect_}
\alias{lim}
\alias{min_}
\alias{max_}
\alias{inf}
\alias{sup}
\alias{bgroup}
\alias{B}
\alias{group}
\alias{G}
\alias{b}
\alias{ceiling_}
\alias{floor_}
\alias{list_}
\alias{paste_}

\alias{alpha}
\alias{beta}
\alias{gamma}
\alias{delta}
\alias{epsilon}
\alias{zeta}
\alias{eta}
\alias{theta}
\alias{iota}
\alias{kappa}
\alias{lambda}
\alias{mu}
\alias{nu}
\alias{xi}
\alias{omicron}
\alias{pi}
\alias{rho}
\alias{sigma}
\alias{tau}
\alias{upsilon}
\alias{phi}
\alias{chi}
\alias{psi}
\alias{omega}

\alias{Alpha}
\alias{Beta}
\alias{Gamma}
\alias{Delta}
\alias{Epsilon}
\alias{Zeta}
\alias{Eta}
\alias{Theta}
\alias{Iota}
\alias{Kappa}
\alias{Lambda}
\alias{Mu}
\alias{Nu}
\alias{Xi}
\alias{Omicron}
\alias{Pi}
\alias{Rho}
\alias{Sigma}
\alias{Tau}
\alias{Upsilon}
\alias{Phi}
\alias{Chi}
\alias{Psi}
\alias{Omega}

\alias{theta1}
\alias{phi1}
\alias{sigma1}
\alias{omega1}

\alias{Upsilon1}
\alias{aleph}
\alias{infinity}
\alias{partialdiff}
\alias{nabla}
\alias{cdots}
\alias{ldots}

\alias{degree}
\alias{minute}
\alias{second}

\alias{ellipsis}
\alias{ellipses}

\alias{degrees}
\alias{minutes}
\alias{seconds}
\title{See \code{?plotmath} for details}
\usage{
plus(x, y)

pos(x)

sum_(x, from = NULL, to = NULL)

bgroup(x, l = "(", r=NULL)
}
\arguments{
\item{x, y}{expressions}

\item{from, to}{operator sub- and super-scripts}

\item{l, r}{left and right bracket definitions}
}
\description{
See \code{?plotmath} for details
}
