\name{mplot}
\alias{mplot}
\title{Render a single plotmath expression}
\usage{
mplot(e, cex = 5, xlim=c(0, 1), ylim=c(0, 1), ...)
}
\arguments{
\item{e}{the expression to render}
\item{cex}{size of the rendered text}
\item{xlim}{plot x limits}
\item{ylim}{plot y limits}
\item{...}{further arguments to text}
}
\description{
Render a single plotmath expression
}
