\name{plot.expression}
\alias{plot.expression}
\title{Plots an expression on the current graphical device.}
\usage{
\method{plot}{expression}(x, ...)
}
\arguments{
\item{x}{A \code{\link{plotmath}} expression.}

\item{...}{Parameters to be passed to the \code{\link{text}} function.}
}
\description{
Plots an expression on the current graphical device.
}

